Move SpeakerPackage to Casts namespace with categories
